requirements.txt
pandas>=2.0.0
pytrends>=4.9.0
matplotlib>=3.7.0
plotly>=5.14.0
python-dotenv>=1.0.0
google-search-results>=2.4.2
